authentication layout construct

+++ backend/controllers/company/companyController.cjs
@@ -1,12 +1,12 @@
 const asyncHandlerSync = require("express-async-handler");
-const { companyModel } = require("../../models/company.models.cjs");
-const Company = companyModel;
+const Company = require("../../models/company.models.cjs");
+const User = require("../../models/user.models.cjs");
 
 //  @desc       Get Company List
 //  @route      Get /api/business/
 //  @access     Private
 const getCompanies = asyncHandlerSync(async (req, res) => {
-  const companies = await Company.find();
+  const companies = await Company.find({ owner: req.user.id });
   res.status(200).json(companies);
 });
 
@@ -14,18 +14,44 @@ const getCompanies = asyncHandlerSync(async (req, res) => {
 //  @route      Get /api/business/new
 //  @access     Private
 const createCompany = asyncHandlerSync(async (req, res) => {
-  const name = req.body.name;
-  const email = req.body.email;
-  const phone = req.body.phone;
-  const about = req.body.about;
+  const { name, url, about, phone, email, logo } = req.body;
 
-  const newCompany = await Company.create({
-    name: name,
+  if (!name || !url || !about || !phone || !email || !logo) {
+    res.status(400);
+    throw new Error("Please fill required input fields");
+  }
+
+  //  Check if company exists
+  const hasCompany = await Company.findOne({ contact: { email } });
+  if (hasCompany) {
+    res.status(400);
+    throw new Error("Company already exists");
+  }
+
+
+  const company = await Company.create({
+    owner: req.user.id,
+    name,
+    url,
+    imgUrl: logo,
     contact: { phone: phone, email: email },
-    about: about,
+    about,
   });
-  if (res.status(200)) console.log("New company data added!");
-  res.status(200).json(newCompany);
+  if (company) {
+    res.status(201).json({
+      _id: company.id,
+      owner: company.owner,
+      name: company.name,
+      url: company.url,
+      logo: company.imgUrl,
+      contact: company.contact,
+      regDate: company.regDate
+    });
+    console.log("New company data added!");
+  } else {
+    res.status(400);
+    throw new Error("Invalid user data");
+  }
 });
 
 //  @desc       Get Company List
@@ -37,17 +63,36 @@ const updateCompany = asyncHandlerSync(async (req, res) => {
     res.status(400);
     throw new Error("Company not found");
   }
-  const name = req.body.name;
-  const email = req.body.email;
-  const phone = req.body.phone;
-  const about = req.body.about;
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (hasCompany.owner.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  const {name, url, imgUrl, about, phone, email} = req.body
   const updateCompany = await Company.findByIdAndUpdate(
     req.params.id,
-    { name: name, contact: { phone: phone, email: email }, about: about },
+    { name, contact: { phone, email }, about, url, imgUrl},
     { new: true }
   );
-  if (res.status(200)) console.log("Company data updated!");
-  res.status(200).json(updateCompany);
+  if (updateCompany) {
+    res.status(201).json({
+      _id: updateCompany.id,
+      owner: updateCompany.owner,
+      name: updateCompany.name,
+      url: updateCompany.url,
+      logo: updateCompany.imgUrl,
+      contact: updateCompany.contact,
+      regDate: updateCompany.regDate,
+    });
+    console.log("Company data updated!");
+  } else {
+    res.status(400);
+    throw new Error("Invalid company data");
+  }
 });
 
 //  @desc       Get Company List
@@ -59,8 +104,17 @@ const deleteCompany = asyncHandlerSync(async (req, res) => {
     res.status(400);
     throw new Error("Company not found");
   }
-  await Company.findByIdAndDelete(req.params.id)
-  res.status(200).json({message: `Company data ${req.params.id} deleted!`})
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (company.owner.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  await Company.findByIdAndDelete(req.params.id);
+  res.status(200).json({ message: `Company(${req.params.id}) data deleted!` });
 });
 
 module.exports = { getCompanies, createCompany, updateCompany, deleteCompany };
